Fix the bug that the release failed.

+++ exchangis-job/exchangis-job-server/src/main/java/com/webank/wedatasphere/exchangis/job/server/restful/external/ExchangisJobDssAppConnRestfulApi.java
@@ -250,7 +250,7 @@ public class ExchangisJobDssAppConnRestfulApi {
         String userName = SecurityFilter.getLoginUsername(request);
         try {
             LOG.info("param: {}", params);
-            if (!hasAuthority(userName, jobInfoService.getJob((Long) params.get("sqoopIds"), true))) {
+            if (!hasAuthority(userName, jobInfoService.getJob(((Integer) params.get("sqoopIds")).longValue(), true))) {
                 return Message.error("You have no permission to import (没有导入权限)");
             }
             response = projectImportServer.importProject(request, params);
@@ -271,7 +271,7 @@ public class ExchangisJobDssAppConnRestfulApi {
         LOG.info("export function params: {}", params);
         Message response = null;
         try {
-            if (!hasAuthority(userName, jobInfoService.getJob((Long) params.get("sqoopIds"), true))) {
+            if (!hasAuthority(userName, jobInfoService.getJob(((Integer) params.get("sqoopIds")).longValue(), true))) {
                 return Message.error("You have no permission to export (没有导出权限)");
             }
             response = jobInfoService.exportProject(params, userName, request);
